# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Banque API Zeynab-Ba",
        "description": "API documentation for Banque application Zeynab-Ba",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8081",
            "description": "API server"
        }
    ],
    "paths": {
...
        "/api/v1/zeynab-ba/comptes": {
            "post": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Créer un nouveau compte",
                "description": "Créer un nouveau compte bancaire",
                "operationId": "createCompte",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "admin_id",
                                    "type",
                                    "devise"
                                ],
                                "properties": {
                                    "admin_id": {
                                        "type": "string",
                                        "format": "uuid",
                                        "example": "550e8400-e29b-41d4-a716-446655440000"
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "cheque",
                                            "courant",
                                            "epargne"
                                        ],
                                        "example": "epargne"
                                    },
                                    "devise": {
                                        "type": "string",
                                        "enum": [
                                            "FCFA",
                                            "EUR",
                                            "USD"
                                        ],
                                        "example": "FCFA"
                                    },
                                    "solde_initial": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 100000
                                    },
                                    "statut": {
                                        "type": "string",
                                        "enum": [
                                            "actif",
                                            "bloque",
                                            "ferme"
                                        ],
                                        "example": "actif"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Compte créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Compte"
                                        },
                                        "_links": {
                                            "properties": {
                                                "self": {
                                                    "properties": {
                                                        "href": {
                                                            "type": "string",
                                                            "example": "/api/v1/zeynab-ba/comptes/{id}"
                                                        },
                                                        "method": {
                                                            "type": "string",
                                                            "example": "GET"
                                                        },
                                                        "rel": {
                                                            "type": "string",
                                                            "example": "self"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Données invalides",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur interne du serveur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": []
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Compte": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "550e8400-e29b-41d4-a716-446655440000"
                    },
                    "numeroCompte": {
                        "type": "string",
                        "example": "C00123456"
                    },
                    "titulaire": {
                        "type": "string",
                        "example": "Amadou Diallo"
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "cheque",
                            "courant",
                            "epargne"
                        ],
                        "example": "epargne"
                    },
                    "solde": {
                        "type": "number",
                        "format": "float",
                        "example": 1250000
                    },
                    "devise": {
                        "type": "string",
                        "enum": [
                            "FCFA",
                            "EUR",
                            "USD"
                        ],
                        "example": "FCFA"
                    },
                    "dateCreation": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-03-15T00:00:00Z"
                    },
                    "statut": {
                        "type": "string",
                        "enum": [
                            "actif",
                            "bloque",
                            "ferme"
                        ],
                        "example": "bloque"
                    },
                    "motifBlocage": {
                        "type": "string",
                        "example": "Inactivité de 30+ jours",
                        "nullable": true
                    },
                    "metadata": {
                        "properties": {
                            "derniereModification": {
                                "type": "string",
                                "format": "date-time",
                                "example": "2023-06-10T14:30:00Z"
                            },
                            "version": {
                                "type": "integer",
                                "example": 1
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
            "ErrorResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Erreur de validation"
                    },
                    "errors": {
                        "type": "object",
                        "nullable": true
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Token d'authentification Bearer JWT"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}